parser - fix the inccorect decision to
When the user double clicks a word,
enventor parser checks whether the word is regarded to some values
that ctxpopup shows the candidates list of (for other values).

In case the user just clicks the ":" and ";"
the parser misjudges it's the values that chould be changes to other values.

Now parser avoids the both cases to do nothing.