editor - optimize line number appending workflow.
Before this patch, it appended each lines number text separately to the line number entry when it reads an edc.
This caused the sizing calculation for every appending.

Now it appends the line numbers to one buffer,
and set it to the line number entry when all lines are readied.

Consequently, the sizing calculation time is much reduced at the edc reading time.