Enventor: Prototipe of Drag-n-Drop feature. Patch 1.
Summary:
This commit contains the basic implementation of Live Edit mode.
With this commit Live Edit toggling is added to Menu, Tools and Ctrl-M hot-key.
Activating of Live Edit mode disables the EDC-Editor and allows user to add new parts by selecting corresponding part type in the menu that appears on right click on the EDJ-Viewer area.

Reviewers: Hermet

Projects: #enventor

Differential Revision: https://phab.enlightenment.org/D1276