EDJ Viewer: fix reseting of layout on edc compiling
Summary:
This commit changes the signal that is used to initiate the reset of edj layout:
using of EIO_MONITOR_FILE_MODIFIED sometimes causes calling of edje_object_file_set() earlier than EDC-file compilation is finished.
So, using of ECORE_EXE_EVENT_DEL to track the end of compilation helps to avoid this problem.

Reviewers: Hermet

Reviewed By: Hermet

Projects: #enventor

Differential Revision: https://phab.enlightenment.org/D1306