edc_parser: Fix to enable inserting templates by live edit.
The problem that template is not inserted by live edit was caused by the wrong
double quotation marks parser.
The double quotation marks are written as "&quote;" in entry markup format and
they are written as "\"" in utf8 format.
Previously, only "&quote;" was used to parse double quotation marks.
Now, "&quote;" is used in entry markup format and "\"" is used in utf8 format
to parse double quotation marks.