Autocomplete: enhance the autocomplete feature.
The hardcoded list of the strings from auto_comp_code.h is replaced by
description of the relationships between keywords.
This descriptions is stored in autocomp.src file.
Now the autocomplete feature is monitoring context of the
inputted text, and provide list of the keywords, that
relevant to current context.
For add new relationships between keywords need
only add it into autocomp.src in correct hierarchical order.