fix the font/live zoom control issue.
When you focus on other window while you pressing the ctrl/alt key at the enventor,
the ctrl/alt key pressed context can be corrupted.

because enventor keeps the extra flag for ctrl/alt key press states.
but in the problem case, the unpress key won't be trigged.

enventor doesn't need to keep the extra flag but track the modifiers in key event.

@fix