Goto: close window from the toolbar.
Summary: moved the focus management  for "goto" window
from base_gui.c inside goto.c.
Added timer, that make delay between unfocus of the "goto" window
and freeing the "goto" internal structures. It is necessary because
when the "goto" window is active and pressed button on toolbar
firstly emitted the signal "unfocused" for window and then
happens button click event.

@fix T1604