fix the part flickering issue.
This problem comes out with a third scenario,
which is, if user uses a ctxpopup slider to change the attributions
of the more than first part state.

In this case, the edj will be reloaded if the attributions are changed.
But still the cursor would be inside of the state,
so the state preview will be switched to the default,
then it turns out a sort of a flickering issue.

@T2969