Redoundo: Support the real-time preview when we undo/redo the code.
When user undo or redo change, the edc file will be recompiled and the
live view will be updated. There are two flow: in case if the change is buildable
(i.e. the edc file was builded sucessfuly, when this change happened)
edc file will be recompiled. In case when change marked as unbilduable -
nothing will happen.

@T3051