Parser: recognize keyword with context.
Uses the autocomplete context parser for recognizing keywords
with the same names in different contexts. For example:
the attributes "min" and "max" for a description context
should be integer values in range from 0 to 1000; but for a
"text" context those attributes should be boolean.

Curently add recognition for a "min" and "max" attributes in the "text" content.

@fix T2712