edc_editor: compare name elaboratley to search the part correctly.
when user press a part on the live view, the part name compareing is triggered.
in that sequence, cursor is jump onto the part wrt clicked part.

since, however, the string compare is not enough
the cursor could go on other part which has a similar name.

now compare them in detail to correct this

@fix T3736