template: don't localize floats in the edje editor
Summary:
Enventor live edit would generate invalid code with
the french locale.
Indeed, french uses a comma (",") as the decimal point,
instead of the dot. With localization enabled, printf()
would automatically localized the floating point,
introducing a comma in the edje code where a dot is
expected...

To avoid this, we now enclose the floats we don't want
to be localized (use the implicit POSIX local: with a dot)
in a function that looks the current locale and replace
the decimal point with a dot.

Test Plan:
(1) Without the patch, call LC_ALL=C enventor and see live
    edit if correct.
(2) Without the patch, call LC_ALL=fr_FR.UTF-8 envetor and
    see live edit generate invalid edje code.
(3) Apply the patch, and repeat steps (1) and (2) and observe
    that both cases now success.

Reviewers: Hermet

Differential Revision: https://phab.enlightenment.org/D4146