edj_viewer: don't update views if they were deactivated.
This fixes the wireframe issue that deactivated views' wireframes are appeared
when the file is updated.

This is only happened if edc collections is consisted with multiple edc files.
If edj file is updated, all generated views will be updated.
and wireframes will be also updated and appeared.

but we don't need to update the deactivated views even if the edj is
updated. we could this with an activated boolean and this will be much
optimized way for many generated views.