re-implement the method for guranteeing the singleton instance on win32.
I found some crash issues that enventor couldn't catch the event.
Have no idea what kinds of the crashes but probably it'd out of the signal
handling.

So re-implement the way of supporting the locking file mechanism
to use the win32 basic File functionality that gurantee removing file
when Enventor is closed in normal or abnormal, either cases.