syntax_indent: Insert spaces by tab key.
Previously, tab key causes that both spaces for indent and spaces for
tab are inserted.
Now, tab key causes that either spaces for indent or spaces for tab are
inserted.

If paragraph is shorter than spaces for indent then only spaces for
indent are inserted. Otherwise, only spaces for tab are inserted.